+++ compiler/internal/codegen/codegen_testmain.go
@@ -24,15 +24,15 @@ func (b *Builder) TestMain(pkg *est.Package, svcs []*est.Service) *File {
 	if pkg.Service != nil {
 		testSvc = pkg.Service.Name
 	}
-	f.Comment("//go:linkname loadApp encore.dev/appruntime/app/appinit.load")
-	f.Func().Id("loadApp").Params().Op("*").Qual("encore.dev/appruntime/app/appinit", "LoadData").Block(
+	f.Comment("//go:linkname loadApp encore.dev/appruntime/app.load")
+	f.Func().Id("loadApp").Params().Op("*").Qual("encore.dev/appruntime/app", "LoadData").Block(
 		Id("static").Op(":=").Op("&").Qual("encore.dev/appruntime/config", "Static").Values(Dict{
 			Id("AuthData"):     b.authDataType(),
 			Id("Testing"):      True(),
 			Id("TestService"):  Lit(testSvc),
 			Id("PubsubTopics"): b.computeStaticPubsubConfig(),
 		}),
-		Return(Op("&").Qual("encore.dev/appruntime/app/appinit", "LoadData").Values(Dict{
+		Return(Op("&").Qual("encore.dev/appruntime/app", "LoadData").Values(Dict{
 			Id("StaticCfg"):   Id("static"),
 			Id("APIHandlers"): Nil(),
 		})),
