+++ compiler/internal/codegen/testdata/TestCodeGenMain__variants.golden
@@ -4,6 +4,7 @@ package main
 import (
 	"encore.app/svc"
 	"encore.dev/appruntime/api"
+	"encore.dev/appruntime/app"
 	"encore.dev/appruntime/app/appinit"
 	"encore.dev/appruntime/config"
 	"reflect"
@@ -11,8 +12,8 @@ import (
 )
 
 // loadApp loads the Encore app runtime.
-//go:linkname loadApp encore.dev/appruntime/app/appinit.load
-func loadApp() *appinit.LoadData {
+//go:linkname loadApp encore.dev/appruntime/app.load
+func loadApp() *app.LoadData {
 	static := &config.Static{
 		AppCommit: config.CommitInfo{
 			Revision:    "",
@@ -38,7 +39,7 @@ func loadApp() *appinit.LoadData {
 		svc.EncoreInternal_ThreeHandler,
 		svc.EncoreInternal_TwoHandler,
 	}
-	return &appinit.LoadData{
+	return &app.LoadData{
 		APIHandlers: handlers,
 		AuthHandler: svc.EncoreInternal_AuthHandlerAuthHandler,
 		StaticCfg:   static,
