+++ runtime/appruntime/app/appinit/appinit.go
@@ -6,9 +6,20 @@ import (
 	"fmt"
 	"os"
 
-	"encore.dev/appruntime/api"
-	"encore.dev/appruntime/app"
-	"encore.dev/appruntime/config"
+	"github.com/rs/zerolog"
+
+	// Note: We should not have any dependencies on any of
+	// the "encore.dev" packages, since we want to use this
+	// package to force initialisation of the Encore runtime to always
+	// occur before any user code can execute. Thus almost all packages inside
+	// encore.dev will import this package - and that would cause a circular dependency.
+
+	// This needs to be imported to allow JSON Iterator and it's dependancies to have
+	// there package level structs loaded, otherwise we get a nil pointer panic.
+	// This is due to the fact that the loadAppInstance() call will initialise
+	// the JSON iterator for the app, but it's package level variables
+	// wouldn't otherwise be setup yet.
+	_ "github.com/json-iterator/go"
 )
 
 // AppMain is the entrypoint to the Encore Application.
@@ -18,33 +29,23 @@ func AppMain() {
 	}
 }
 
-// singleton is the instance of the Encore app.
-var singleton *app.App
-
-// load is provided by the code-generated main package
-// and linked here using go:linkname.
-func load() *LoadData
-
-type LoadData struct {
-	StaticCfg   *config.Static
-	APIHandlers []api.Handler
-	AuthHandler api.AuthHandler
+type AppInstance interface {
+	Run() error
+	RootLogger() *zerolog.Logger
+	GetSecret(key string) (string, bool)
 }
 
-// We load everything during init so that the whole runtime is available to the Encore app
-// even from within the app's init functions. The AppMain function runs later.
+// singleton is the instance of the Encore app.
+var singleton AppInstance
+
+// loadAppInstance is provided by the "appruntime/app"
+// and linked here using go:linkname. This indirection is
+// done to avoid a dependency on the "appruntime/app" package which
+// would create an circular dependency on the whole runtime.
+func loadAppInstance() AppInstance
+
 func init() {
-	data := load()
-	cfg := &config.Config{
-		Runtime: config.ParseRuntime(config.GetAndClearEnv("ENCORE_RUNTIME_CONFIG")),
-		Secrets: config.ParseSecrets(config.GetAndClearEnv("ENCORE_APP_SECRETS")),
-		Static:  data.StaticCfg,
-	}
-	singleton = app.New(&app.NewParams{
-		Cfg:         cfg,
-		APIHandlers: data.APIHandlers,
-		AuthHandler: data.AuthHandler,
-	})
+	singleton = loadAppInstance()
 }
 
 // LoadSecret loads the secret with the given key.
