+++ runtime/appruntime/testsupport/runtimehooks_app.go
@@ -5,6 +5,8 @@ package testsupport
 import (
 	"testing"
 	_ "unsafe" // for go:linkname
+
+	_ "encore.dev/appruntime/app/appinit" // Force the app to initialise all singletons before these functions can be used
 )
 
 var Singleton *Manager
