+++ runtime/beta/auth/pkgfn.go
@@ -2,6 +2,8 @@
 
 package auth
 
+import _ "encore.dev/appruntime/app/appinit" // Force the app to initialise all singletons before these functions can be used
+
 //publicapigen:drop
 var Singleton *Manager // injected on app init
